+++ src/client.rs
@@ -1,21 +1,23 @@
-use crate::discovery::{self, Config, Discovered};
-use crate::error::{
-    ClientError, Decode, Error, Expiry, Jose, Mismatch, Missing, Userinfo as ErrorUserinfo,
-    Validation,
+use crate::{
+    discovered,
+    error::{
+        ClientError, Decode, Error, Expiry, Jose, Mismatch, Missing, Userinfo as ErrorUserinfo,
+        Validation,
+    },
+    Bearer, Claims, Config, Discovered, IdToken, OAuth2Error, Options, Provider, StandardClaims,
+    Token, Userinfo,
 };
-use crate::{Bearer, Claims, IdToken, OAuth2Error, Provider, StandardClaims, Token};
-use biscuit::jwa::{self, SignatureAlgorithm};
-use biscuit::jwk::{AlgorithmParameters, JWKSet};
-use biscuit::jws::{Compact, Secret};
-use biscuit::{CompactJson, Empty, SingleOrMultiple};
-use chrono::{Duration, NaiveDate, Utc};
+use biscuit::{
+    jwa::{self, SignatureAlgorithm},
+    jwk::{AlgorithmParameters, JWKSet},
+    jws::{Compact, Secret},
+    CompactJson, Empty, SingleOrMultiple,
+};
+use chrono::{Duration, Utc};
 use reqwest::header::{ACCEPT, CONTENT_TYPE};
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::marker::PhantomData;
 use url::{form_urlencoded::Serializer, Url};
-use validator::Validate;
-use validator_derive::Validate;
 
 /// OAuth 2.0 client.
 #[derive(Debug)]
@@ -58,8 +60,8 @@ impl<C: CompactJson + Claims> Client<Discovered, C> {
         issuer: Url,
     ) -> Result<Self, Error> {
         let http_client = reqwest::Client::new();
-        let config = discovery::discover(&http_client, issuer).await?;
-        let jwks = discovery::jwks(&http_client, config.jwks_uri.clone()).await?;
+        let config = discovered::discover(&http_client, issuer).await?;
+        let jwks = discovered::jwks(&http_client, config.jwks_uri.clone()).await?;
         let provider = Discovered(config);
         Ok(Self::new(
             provider,
@@ -352,163 +354,6 @@ impl<C: CompactJson + Claims> Client<Discovered, C> {
             None => Err(ErrorUserinfo::NoUrl.into()),
         }
     }
-}
-
-/// Optional parameters that [OpenID specifies](https://openid.net/specs/openid-connect-basic-1_0.html#RequestParameters) for the auth URI.
-/// Derives Default, so remember to ..Default::default() after you specify what you want.
-#[derive(Default)]
-pub struct Options {
-    /// MUST contain openid. By default this is ONLY openid. Official optional scopes are
-    /// email, profile, address, phone, offline_access. Check the Discovery config
-    /// `scopes_supported` to see what is available at your provider!
-    pub scope: Option<String>,
-    pub state: Option<String>,
-    pub nonce: Option<String>,
-    pub display: Option<Display>,
-    pub prompt: Option<std::collections::HashSet<Prompt>>,
-    pub max_age: Option<Duration>,
-    pub ui_locales: Option<String>,
-    pub claims_locales: Option<String>,
-    pub id_token_hint: Option<String>,
-    pub login_hint: Option<String>,
-    pub acr_values: Option<String>,
-}
-
-/// The userinfo struct contains all possible userinfo fields regardless of scope. [See spec.](https://openid.net/specs/openid-connect-basic-1_0.html#StandardClaims)
-// TODO is there a way to use claims_supported in config to simplify this struct?
-#[derive(Debug, Deserialize, Serialize, Validate)]
-pub struct Userinfo {
-    #[serde(default)]
-    /// Subject - Identifier for the End-User at the Issuer.
-    pub sub: Option<String>,
-    #[serde(default)]
-    /// End-User's full name in displayable form including all name parts, possibly including titles and suffixes, ordered according to the End-User's locale and preferences.
-    pub name: Option<String>,
-    #[serde(default)]
-    /// Given name(s) or first name(s) of the End-User. Note that in some cultures, people can have multiple given names; all can be present, with the names being separated by space characters.
-    pub given_name: Option<String>,
-    #[serde(default)]
-    /// Surname(s) or last name(s) of the End-User. Note that in some cultures, people can have multiple family names or no family name; all can be present, with the names being separated by space characters.
-    pub family_name: Option<String>,
-    #[serde(default)]
-    /// Middle name(s) of the End-User. Note that in some cultures, people can have multiple middle names; all can be present, with the names being separated by space characters. Also note that in some cultures, middle names are not used.
-    pub middle_name: Option<String>,
-    #[serde(default)]
-    /// Casual name of the End-User that may or may not be the same as the given_name. For instance, a nickname value of Mike might be returned alongside a given_name value of Michael.
-    pub nickname: Option<String>,
-    #[serde(default)]
-    /// Shorthand name by which the End-User wishes to be referred to at the RP, such as janedoe or j.doe. This value MAY be any valid JSON string including special characters such as @, /, or whitespace. The RP MUST NOT rely upon this value being unique, as discussed in Section 5.7.
-    pub preferred_username: Option<String>,
-    #[serde(default)]
-    /// URL of the End-User's profile page. The contents of this Web page SHOULD be about the End-User.
-    pub profile: Option<Url>,
-    #[serde(default)]
-    /// URL of the End-User's profile picture. This URL MUST refer to an image file (for example, a PNG, JPEG, or GIF image file), rather than to a Web page containing an image. Note that this URL SHOULD specifically reference a profile photo of the End-User suitable for displaying when describing the End-User, rather than an arbitrary photo taken by the End-User.
-    pub picture: Option<Url>,
-    #[serde(default)]
-    /// URL of the End-User's Web page or blog. This Web page SHOULD contain information published by the End-User or an organization that the End-User is affiliated with.
-    pub website: Option<Url>,
-    #[serde(default)]
-    #[validate(email)]
-    /// End-User's preferred e-mail address. Its value MUST conform to the RFC 5322 [RFC5322] addr-spec syntax. The RP MUST NOT rely upon this value being unique, as discussed in Section 5.7.
-    pub email: Option<String>,
-    #[serde(default)]
-    /// True if the End-User's e-mail address has been verified; otherwise false. When this Claim Value is true, this means that the OP took affirmative steps to ensure that this e-mail address was controlled by the End-User at the time the verification was performed. The means by which an e-mail address is verified is context-specific, and dependent upon the trust framework or contractual agreements within which the parties are operating.
-    pub email_verified: bool,
-    // Isn't required to be just male or female
-    #[serde(default)]
-    /// End-User's gender. Values defined by this specification are female and male. Other values MAY be used when neither of the defined values are applicable.
-    pub gender: Option<String>,
-    // ISO 9601:2004 YYYY-MM-DD or YYYY.
-    #[serde(default)]
-    /// End-User's birthday, represented as an ISO 8601:2004 [ISO8601‑2004] YYYY-MM-DD format. The year MAY be 0000, indicating that it is omitted. To represent only the year, YYYY format is allowed. Note that depending on the underlying platform's date related function, providing just year can result in varying month and day, so the implementers need to take this factor into account to correctly process the dates.
-    pub birthdate: Option<NaiveDate>,
-    // Region/City codes. Should also have a more concrete serializer form.
-    /// String from zoneinfo [zoneinfo] time zone database representing the End-User's time zone. For example, Europe/Paris or America/Los_Angeles.
-    #[serde(default)]
-    pub zoneinfo: Option<String>,
-    // Usually RFC5646 langcode-countrycode, maybe with a _ sep, could be arbitrary
-    #[serde(default)]
-    /// End-User's locale, represented as a BCP47 [RFC5646] language tag. This is typically an ISO 639-1 Alpha-2 [ISO639‑1] language code in lowercase and an ISO 3166-1 Alpha-2 [ISO3166‑1] country code in uppercase, separated by a dash. For example, en-US or fr-CA. As a compatibility note, some implementations have used an underscore as the separator rather than a dash, for example, en_US; Relying Parties MAY choose to accept this locale syntax as well.
-    pub locale: Option<String>,
-    // Usually E.164 format number
-    #[serde(default)]
-    /// End-User's preferred telephone number. E.164 [E.164] is RECOMMENDED as the format of this Claim, for example, +1 (425) 555-1212 or +56 (2) 687 2400. If the phone number contains an extension, it is RECOMMENDED that the extension be represented using the RFC 3966 [RFC3966] extension syntax, for example, +1 (604) 555-1234;ext=5678.
-    pub phone_number: Option<String>,
-    #[serde(default)]
-    /// True if the End-User's phone number has been verified; otherwise false. When this Claim Value is true, this means that the OP took affirmative steps to ensure that this phone number was controlled by the End-User at the time the verification was performed. The means by which a phone number is verified is context-specific, and dependent upon the trust framework or contractual agreements within which the parties are operating. When true, the phone_number Claim MUST be in E.164 format and any extensions MUST be represented in RFC 3966 format.
-    pub phone_number_verified: bool,
-    #[serde(default)]
-    /// End-User's preferred postal address. The value of the address member is a JSON [RFC4627] structure containing some or all of the members defined in Section 5.1.1.
-    pub address: Option<Address>,
-    #[serde(default)]
-    /// Time the End-User's information was last updated. Its value is a JSON number representing the number of seconds from 1970-01-01T0:0:0Z as measured in UTC until the date/time.
-    pub updated_at: Option<i64>,
-}
-
-/// The four values for the preferred display parameter in the Options. See spec for details.
-pub enum Display {
-    Page,
-    Popup,
-    Touch,
-    Wap,
-}
-
-impl Display {
-    fn as_str(&self) -> &'static str {
-        use self::Display::*;
-        match *self {
-            Page => "page",
-            Popup => "popup",
-            Touch => "touch",
-            Wap => "wap",
-        }
-    }
-}
-
-/// The four possible values for the prompt parameter set in Options. See spec for details.
-#[derive(PartialEq, Eq, Hash)]
-pub enum Prompt {
-    None,
-    Login,
-    Consent,
-    SelectAccount,
-}
-
-impl Prompt {
-    fn as_str(&self) -> &'static str {
-        use self::Prompt::*;
-        match *self {
-            None => "none",
-            Login => "login",
-            Consent => "consent",
-            SelectAccount => "select_account",
-        }
-    }
-}
-
-/// Address Claim struct. Can be only formatted, only the rest, or both.
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Address {
-    #[serde(default)]
-    /// Full mailing address, formatted for display or use on a mailing label. This field MAY contain multiple lines, separated by newlines. Newlines can be represented either as a carriage return/line feed pair ("\r\n") or as a single line feed character ("\n").
-    pub formatted: Option<String>,
-    #[serde(default)]
-    /// Full street address component, which MAY include house number, street name, Post Office Box, and multi-line extended street address information. This field MAY contain multiple lines, separated by newlines. Newlines can be represented either as a carriage return/line feed pair ("\r\n") or as a single line feed character ("\n").
-    pub street_address: Option<String>,
-    #[serde(default)]
-    /// City or locality component.
-    pub locality: Option<String>,
-    #[serde(default)]
-    /// State, province, prefecture, or region component.
-    pub region: Option<String>,
-    // Countries like the UK use alphanumeric postal codes, so you can't just use a number here
-    #[serde(default)]
-    /// Zip code or postal code component.
-    pub postal_code: Option<String>,
-    #[serde(default)]
-    /// Country name component.
-    pub country: Option<String>,
 }
 
 impl<P, C> Client<P, C>
